requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
pydantic[email]==2.9.2
jinja2==3.1.4
requests==2.32.3
pandas==2.2.2
reportlab==4.2.5
pytest==8.3.3
httpx==0.27.2
